requirements.txt
pandas
numpy
matplotlib
streamlit
pathlib
ensure
html5lib
PyYaml
python-box
bs4
requests

-e .